requirements.txt
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
cloudinary==1.44.1
colorama==0.4.6
dnspython==2.8.0
Flask==3.1.3
gunicorn==25.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
pillow==12.1.1
pymongo==4.16.0
PyQRCode==1.2.1
python-dotenv==1.2.1
reportlab==4.4.10
six==1.17.0
urllib3==2.6.3
Werkzeug==3.1.6
